+++ crud/insert_many.lua
@@ -54,7 +54,7 @@ local function insert_many_on_storage(space_name, tuples, opts)
         bucket_ids[tuple[utils.get_bucket_id_fieldno(space)]] = true
     end
 
-    local ref_ok, bucket_ref_err, unref = bucket_ref_unref.bucket_refrw_many(bucket_ids)
+    local ref_ok, bucket_ref_err, unref = bucket_ref_unref.bucket_refrw_many(bucket_ids, space.engine)
     if not ref_ok then
         return nil, bucket_ref_err
     end
@@ -100,7 +100,7 @@ local function insert_many_on_storage(space_name, tuples, opts)
                 end
 
                 if opts.rollback_on_error == true then
-                    local unref_ok, bucket_unref_err = unref(bucket_ids)
+                    local unref_ok, bucket_unref_err = unref(bucket_ids, space.engine)
                     box.rollback()
                     if not unref_ok then
                         return nil, bucket_unref_err
@@ -113,7 +113,7 @@ local function insert_many_on_storage(space_name, tuples, opts)
                     return nil, errs, replica_schema_version
                 end
 
-                local unref_ok, bucket_unref_err = unref(bucket_ids)
+                local unref_ok, bucket_unref_err = unref(bucket_ids, space.engine)
                 box.commit()
                 if not unref_ok then
                     return nil, bucket_unref_err
@@ -128,7 +128,7 @@ local function insert_many_on_storage(space_name, tuples, opts)
 
     if next(errs) ~= nil then
         if opts.rollback_on_error == true then
-            local unref_ok, bucket_unref_err = unref(bucket_ids)
+            local unref_ok, bucket_unref_err = unref(bucket_ids, space.engine)
             box.rollback()
             if not unref_ok then
                 return nil, bucket_unref_err
@@ -141,7 +141,7 @@ local function insert_many_on_storage(space_name, tuples, opts)
             return nil, errs, replica_schema_version
         end
 
-        local unref_ok, bucket_unref_err = unref(bucket_ids)
+        local unref_ok, bucket_unref_err = unref(bucket_ids, space.engine)
         box.commit()
         if not unref_ok then
             return nil, bucket_unref_err
@@ -150,7 +150,7 @@ local function insert_many_on_storage(space_name, tuples, opts)
         return inserted_tuples, errs, replica_schema_version
     end
 
-    local unref_ok, bucket_unref_err = unref(bucket_ids)
+    local unref_ok, bucket_unref_err = unref(bucket_ids, space.engine)
     box.commit()
     if not unref_ok then
         return nil, bucket_unref_err