+++ crud/replace.lua
@@ -44,7 +44,7 @@ local function replace_on_storage(space_name, tuple, opts)
     end
 
     local bucket_id = tuple[utils.get_bucket_id_fieldno(space)]
-    local ref_ok, bucket_ref_err, unref = bucket_ref_unref.bucket_refrw(bucket_id)
+    local ref_ok, bucket_ref_err, unref = bucket_ref_unref.bucket_refrw(bucket_id, space.engine)
     if not ref_ok then
         return nil, bucket_ref_err
     end
@@ -59,7 +59,7 @@ local function replace_on_storage(space_name, tuple, opts)
         fetch_latest_metadata = opts.fetch_latest_metadata,
     }, space, tuple)
 
-    local unref_ok, err_unref = unref(bucket_id)
+    local unref_ok, err_unref = unref(bucket_id, space.engine)
     if not unref_ok then
         return nil, err_unref
     end