+++ crud/replace_many.lua
@@ -65,7 +65,7 @@ local function replace_many_on_storage(space_name, tuples, opts)
         bucket_ids[tuple[utils.get_bucket_id_fieldno(space)]] = true
     end
 
-    local ref_ok, bucket_ref_err, unref = bucket_ref_unref.bucket_refrw_many(bucket_ids)
+    local ref_ok, bucket_ref_err, unref = bucket_ref_unref.bucket_refrw_many(bucket_ids, space.engine)
     if not ref_ok then
         return nil, bucket_ref_err
     end
@@ -103,7 +103,7 @@ local function replace_many_on_storage(space_name, tuples, opts)
                 end
 
                 if opts.rollback_on_error == true then
-                    local unref_ok, bucket_unref_err = unref(bucket_ids)
+                    local unref_ok, bucket_unref_err = unref(bucket_ids, space.engine)
                     box.rollback()
                     if not unref_ok then
                         return nil, bucket_unref_err
@@ -115,7 +115,7 @@ local function replace_many_on_storage(space_name, tuples, opts)
                     return nil, errs, replica_schema_version
                 end
 
-                local unref_ok, bucket_unref_err = unref(bucket_ids)
+                local unref_ok, bucket_unref_err = unref(bucket_ids, space.engine)
                 box.commit()
                 if not unref_ok then
                     return nil, bucket_unref_err
@@ -129,7 +129,7 @@ local function replace_many_on_storage(space_name, tuples, opts)
 
     if next(errs) ~= nil then
         if opts.rollback_on_error == true then
-            local unref_ok, bucket_unref_err = unref(bucket_ids)
+            local unref_ok, bucket_unref_err = unref(bucket_ids, space.engine)
             box.rollback()
             if not unref_ok then
                 return nil, bucket_unref_err
@@ -142,7 +142,7 @@ local function replace_many_on_storage(space_name, tuples, opts)
             return nil, errs, replica_schema_version
         end
 
-        local unref_ok, bucket_unref_err = unref(bucket_ids)
+        local unref_ok, bucket_unref_err = unref(bucket_ids, space.engine)
         box.commit()
         if not unref_ok then
             return nil, bucket_unref_err
@@ -151,7 +151,7 @@ local function replace_many_on_storage(space_name, tuples, opts)
         return inserted_tuples, errs, replica_schema_version
     end
 
-    local unref_ok, bucket_unref_err = unref(bucket_ids)
+    local unref_ok, bucket_unref_err = unref(bucket_ids, space.engine)
     box.commit()
     if not unref_ok then
         return nil, bucket_unref_err